requirements.txt
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==0.5.0
Django==3.2.12
django-ckeditor==6.2.0
django-cors-headers==3.10.1
django-heroku==0.3.1
django-js-asset==2.0.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
openapi-codec==1.3.2
psycopg2==2.9.3
PyJWT==2.3.0
pytz==2022.1
requests==2.27.1
simplejson==3.17.6
sqlparse==0.4.2
twilio==7.7.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==5.3.0
